+++ ContentPatcher/Framework/Conditions/ConditionType.cs
@@ -188,6 +188,9 @@ namespace ContentPatcher.Framework.Conditions
         /****
         ** Metadata
         ****/
+        /// <summary>A token which returns the absolute path for a file in the content pack's folder.</summary>
+        AbsoluteFilePath,
+
         /// <summary>A token which returns the first input argument which matches an existing file in the content pack.</summary>
         FirstValidFile,