+++ ContentPatcher/docs/release-notes.md
@@ -15,6 +15,8 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
   * You can now edit arbitrary data models (e.g. non-list/dictionary assets added by mods).
   * You can now edit list entries by their index position, even if they have no ID available.
 * Added [token aliases](author-guide/tokens.md#aliases) (thanks to Shockah!).
+* Added new tokens:
+  * [`AbsoluteFilePath`](author-guide/tokens.md#AbsoluteFilePath) to get the full path for a file in the content pack folder.
 * Added [`AnyPlayer` option](author-guide/tokens.md#target-player) for per-player tokens.
 * The `Format` field now ignores the third number, so `1.24.3` is equivalent to `1.24.0` instead of showing an error.
 * The deprecated `Enabled` patch field is now obsolete.